Fix broken links in generated changelog
There were two problems here:

One is that I didn't pass a "v" prefix when I specified the version
number (which is a very easy mistake to make, so we should probably add
some kind of safeguard against that or perhaps even an auto-fix).

This led to some broken links on GitHub, where the tags all have a "v"
prefix; but the other problem was that we weren't properly stripping the
possible ".git" suffix in the remote URL, which was also breaking links.
There was code in there to do the stripping, but due to the vagaries of
regular expressions, the `\S+` greedy match ended up slurping up the
".git" suffix and including it, and turning it into a non-greedy match
would just make it consume only an incomplete URL. Maybe you could do
this with some clever combination of lookahead and/or lookbehind, but I
don't think it would be very readable, so I went with the dumb
`replace()` based approach instead.

Regenerated the changelog by resetting it to its former state and then
running:

    scripts/changelog.js --version=v9.0.0-alpha.1 --to=v9.0.0-alpha.1