Teach the changelog generator to correct errors with "v" prefix
When I used this earlier today I forgot to use a "v" prefix when I
specified the version number. This change makes the tool a little more
foolproof by warning you if you forget, and adding it on for you (you
can disable the autocorrection and force it to do what you say anyway
with `--force`).

In this repo we always use the "v" prefix but we might not in others,
and I am anticipating using this script in other places, so instead of
hard-coding a check for a "v" prefix, what this actually does is look at
the existing tags and decide whether "v" is used by convention. If it is
not, it actually behaves in the inverse manner: it warns about an
unexpected "v" prefix and automatically removes it, unless you use
`--force`.