Fix "gulp upgrade" task for 6.2 themes (#244)
I had mistakenly removed the gulp-prompt dependency in bea3c7035673e80
because I thought it was unused and the old version (0.1.2) was adding
noise to the `yarn audit` report.

Turns out it still is used by the 6.2 "upgrade" task, but the usage
didn't show up in my search because there is no explicit `require()`
call, only dynamic loading via the "gulp-load-plugins" package. (And now
that I look at the list of plugins in the theme tasks package.json, I
see some others that I suspect may not be used anywhere, but auditing
those must be a task for a different commit.)

Here we add it back, but now a much newer version (1.2.0). The `yarn
audit` output is unaffected.

Test plan: `gulp upgrade` in a 6.2 theme; `yarn test`.

Builds on: https://github.com/liferay/liferay-js-themes-toolkit/pull/247

Closes: https://github.com/liferay/liferay-js-themes-toolkit/issues/244