Change merge order so that local overrides have precedence
Noticed while preparing #37 that if I made a local override in my
"package.json" it had no effect. That's because liferay-npm-scripts is
taking whatever user settings are provided (either in "package.json" or
in a "jest.config.js" file) and then overwriting them with the top-level
defaults defined in its own "config/jest.json" file.

This seems to be the wrong behavior for defaults, which should only take
effect in the absence of a more explicit setting.

Before merging this, we should check the portal repo to make sure we're
not relying on the old behavior anywhere; at the very least we can clean
out some settings from the repo that can't possibly be having any
effect.