feat: Provide hint to update to v9.x in order to upgrade a 7.1 theme (#257)
Test plan:

In a 7.1 theme gulpfile, change the `require` to point at this PR (ie.
`require('../packages/liferay-theme-tasks')`.

Then, run `gulp upgrade` in the 7.1 theme and see:

    Error: Task 'upgrade' is not supported for themes with
    version '7.1' in this version of 'liferay-theme-tasks' -
    to upgrade a 7.1 theme please run "npm install --save-dev
    liferay-theme-tasks@9.0.0-alpha.1" and then "gulp upgrade" again

Run `npm install --save-dev liferay-theme-tasks@9.0.0-alpha.1`. Change
the `require` back to `require('liferay-theme-tasks')` so that the theme
will use the code we just told the user to download, then run
`gulp upgrade` again and see it work.

Note that users obviously won't have to edit their gulpfiles; it's only
necessary for testing in a dev environment like this one.

Closes: https://github.com/liferay/liferay-js-themes-toolkit/issues/257