Kill unused Lerna set-up and switch to Yarn workspaces (#42)
We're not actually using Lerna in this repo despite the config, so this
commit gets rid of it. Instead, let's use Yarn workspaces, which yields
a nice reduction in lockfile size:

10 files changed, 8043 insertions(+), 22085 deletions(-)

Test plan: `yarn test` in "packages/liferay-jest-junit-reporter", which
is the only package we actually have tests for in this repo. Note that
the snapshot had a hardcoded "bryceosterhaus" in it, which failed
because it doesn't match my username on this machine ("greghurrell"), so
I tweaked the test to avoid using user-specific information, and I
updated the snapshot accordingly.

Related: https://github.com/liferay/liferay-npm-tools/issues/42