chore: remove backup/revert functionality (9.x) (#267)
This is the 9.x cherry-pick of the corresponding change for 8.x:

    https://github.com/liferay/liferay-js-themes-toolkit/pull/269

We have functionality that creates a backup of files before attempting
an upgrade and then offer a revert functionality in the case that an
error is thrown.

In practice, this isn't as good as proper version control because:

- You cannot "double revert" because we only make a backup of the most
  recent change.
- In the event that an upgrade fails we blow away the upgraded files by
  doing an automatic revert, preventing the user from inspecting them.
- We incur complexity in the codebase, opening the door to bugs.
- We can't provide a very nice upgrade-and-revert experience moving from
  6.2 to 7.0 to 7.1 to 7.2 and back because those moves involve switching
  versions of the toolkit from v8 to v9 along the way, as described in
  #267.

So this commit removes the backup and revert functionality and instead
shows a prompt that advises people to make a backup before upgrading,
and confirming their intent to proceed: in practice, I expect that
developers using the toolkit will all have their code in Git or an
equivalent version control system, so they *already* will have a backup.

If an error occurs, we now show a message stating that the upgraded
theme files have been left in place for inspection.

Test plan:

    # Use the generator to create a new 7.1 theme:
    npm install generator-liferay-theme
    yo ./node_modules/generator-liferay-theme

    # Update the `require('liferay-theme-tasks')` in the theme gulpfile
    # to point at the version of the code that I am testing
    cd my-liferay-theme
    edit gulpfile

    # Run the upgrade:
    gulp upgrade

See:

    [17:16:33] Starting 'upgrade'...
    ? We recommend creating a backup of your theme files before proceeding. Are you sure you wish to start the upgrade proc
    ess? No
    [17:16:43] Finished 'upgrade' after 10 s

Run it again and say "y" this time; see:

    [17:16:58] Starting 'upgrade'...
    ? We recommend creating a backup of your theme files before proceeding. Are you sure you wish to start the upgrade proc
    ess? Yes
    [17:16:59] Starting 'upgrade:config'...
    [17:16:59] Finished 'upgrade:config' after 24 ms
    [17:16:59] Starting 'upgrade:dependencies'...
    [17:17:16] Finished 'upgrade:dependencies' after 17 s
    [17:17:16] Finished 'upgrade' after 18 s

Now hardcode the upgrade task to throw an error and see:

    [17:21:19] Finished 'upgrade:dependencies' after 7.4 s
    [17:21:19] Error: something went wrong during the upgrade task, leaving the theme files in place for inspection.
    [17:21:19] Error: Something blew up!

Closes: https://github.com/liferay/liferay-js-themes-toolkit/issues/267