fix: detect git@github.com remotes in addition to HTTPs remotes
This will stop us from making double updates to the CHANGELOG like we
did here:

https://github.com/liferay/alloy-editor/commit/0b8145051ea58e4c0

The cause was: we'd run, fail to detect the `git@github` remote, but
still update the changelog, and then run again with `--remote`, and
again update it.

The fix is to try harder to detect the remote.

A deeper fix may be to `error` instead of `warn` when we detect a
duplicate heading.