chore: Update references to "DXP" in user-facing messages (8.x) (#280)
On the 8.x branch, the main place we refer to "DXP" is in the README,
so I changed it to say "DXP or Portal CE". Repeating that name in the
capabilities table would have been very cumbersome, so I stripped out
the names entirely, making them implicit.

The CHANGELOG also has a couple of references, so I updated those too.
Because we're not using `github_changelog_generator` any more, we don't
have to worry about the edits getting overwritten.

Related: https://github.com/liferay/liferay-js-themes-toolkit/issues/280