feat: add "theme" subcommand (#64)
With this, we can replace scripts like theses ones in liferay-portal:

    "build": "gulp build --css-common-path ./build_gradle/frontend-css-common --styled-path ../../frontend-theme/frontend-theme-styled/src/main/resources/META-INF/resources/_styled --unstyled-path ../../frontend-theme/frontend-theme-unstyled/src/main/resources/META-INF/resources/_unstyled"

with:

    "build": "liferay-npm-scripts theme build"

I added a bunch of tests because this involves traversing the filesystem
and I didn't want to screw it up. Using runtime traversal rather than
hardcoding relative paths in order to make us _somewhat_ resilient to
possible future changes in the structure of the filesystem layout. As it
is right now, we have two variants in liferay-portal, which I have
reflected in the structure of the fixtures:

- Marketplace themes like fjord that live in modules like
  "frontend-theme-fjord/frontend-theme-fjord".
- Core themes which like "classic" that live under "frontend-theme",
  right next to the "unstyled" and "styled" dependencies.

Test plan: `yarn test`

Closes: https://github.com/liferay/liferay-npm-tools/issues/64