feat: add top-level shortcut to run all tests (#67)
This it the simplest thing I could think of to get tests running from
the top-level. Ideally though, we could just run jest once at the top.
That would be pretty easy to set-up, but would require a bit of extra
work to make it happen while still preserving the ability to run the
tests at the level of the individual packages, so I am keeping it simple
for now.

Test plan: `yarn test` at top level and see all run. Then `yarn test` at
individual package level to prove that that still works too.

Closes: https://github.com/liferay/liferay-npm-tools/issues/67