feat: add ability to config PostCSS plug-ins (#126)
Prior to this commit, we would look at the plugins named in the theme's
"package.json" file under "liferayTheme" -> "postcss" and load them.
"autoprefixer" was whitelisted because we bundle it with
liferay-theme-tasks (ie. we would just do `require('autoprefixer')`).

For the others we would go through our `resolveDependency` helper, which
actually seems to be a pointless in the context of PostCSS. I'm going to
remove that special casing in the next commit, in order to make it
easier to review.

With this commit, we allow for configuration by any of the mechanisms
supported by postcss-load-config:

https://www.npmjs.com/package/postcss-load-config

(gulp-postcss will use postcss-load-config transparently behind the
scenes).

What this means is that you can add a "postcss" property to you
"package.json" like this:

```
	"postcss": {
		"plugins": {
			"autoprefixer": {
				"grid": "autoplace"
			}
		}
	},
```

or, you can create one of:

- .postcssrc
- .postcssrc.js
- .postcssrc.json
- .postcssrc.yml
- postcss.config.js

If you try to supply a PostCSS config by one of these means *and* also
via "liferayTheme" -> "postcss", we throw an error.

Test plan:

- Confirm that old style of config still works by building
  liferay-theme-fjord with "liferayTheme" -> "postcss" ->
  `["autoprefixer"]`.
- Add a .postcssrc file and see this error:

    PostCSS must be configured in only one place but it was configured
    in both "liferayTheme" properties of the package.json and also in
    .postcssrc

- Remove that file and set a "postcss" config in the "package.json"
  instead and see this error:

    PostCSS must be configured in only one place but it was configured
    in both "liferayTheme" properties of the package.json and also in
    package.json "postcss"

- Remove the duplicate config from the "liferayTheme" config and see the
  build work.

I would have liked to add automated tests for this, but it is basically
impractical to do anything above what we're already doing in the Jest
test suite (ie. running the build and seeing the hook callbacks fired).
So for now I am just going with manual QA.

Closes: https://github.com/liferay/liferay-js-themes-toolkit/issues/126