chore: remove use of `resolveDependency()` in PostCSS handler (#126)
In preparing the previous commit I looked at why we were calling
`resolveDependency()` on the plug-in names passed in via "liferayTheme"
-> "postcss".

That function does two things:

1. It handles overrides provided by environment variables or
   command-line switches for liferay-frontend-common-css,
   liferay-frontend-theme-styled, and liferay-frontend-theme-unstyled.
   Obviously, this has nothing to do with loading PostCSS plug-ins.
2. It calls `resolve.sync()` relative to the current working directory
  (eg. yielding an absolute path to the local
  `node_modules/autoprefixer/lib/autoprefixer.js` or similar); then it
  passes *that* to `require.resolve()`, which looks like a no-op, and
  finally strips off the filename and returns the dirname
  (ie. absolute path to `node_modules/autoprefixer/lib`).

Now, that is only going to work if the plug-in happens to have a file
called "index.js" in that directory; that's not the case for
autoprefixer, but our special casing means that we never call
`resolveDependency()` for it anyway. I suspect that it is quite probable
that that isn't true for a whole bunch of postcss plug-ins. Two that I
sampled at random don't, for example, both having an "index.cjs.js"
instead:

- https://github.com/postcss/postcss-color-hex-alpha
- https://github.com/postcss/postcss-custom-media

So, my belief is that this was at best unnecessary and at worst broken,
and I am pretty sure it was broken.

Test plan: As per previous commit.

Related: https://github.com/liferay/liferay-js-themes-toolkit/issues/126