chore: get rid of "resolve" dependency (#126)
As I mentioned in the last commit, calling `resolve.sync` relative to
the current working directory _looks_ like a no-op, identical to calling
`require.resolve`, so I don't think we need this dependency any more.

In practice, I think it _is_ a no-op most of the time, because we don't
go around `chdir`-ing from one place to another during actual builds,
only in the test suite.

If you read the module resolution spec, you see that the starting point
from which we search for node modules is determined by the file that the
`require` appears in, even if it is an unqualified one like
`require('autoprefixer')`. That is, if you `process.chdir()` like we do
in the test suite, then your `require()` won'd look in the local
"node_modules" folder, so my initial naïve attempt at removing this
dependency caused one of the tests to fail.

https://nodejs.org/api/modules.html

Maybe that means the test is no good, as I don't think there's any real
possibility of a failure in real usage, but just to be safe, I am taking
the cautious route and replacing the unwanted dependency with built-in
functionality that has existed since Node v8.9.0.

Test plan: `yarn test`, and for good measure, repeat the test plan from
the previous two commits, which involves building and deploying a theme
in the context of the liferay-portal repo.

Related: https://github.com/liferay/liferay-js-themes-toolkit/issues/126

fixup! chore: get rid of "resolve" dependency (#126)