chore: get rid of "resolve" dependency (#126)
This is the 8.x cherry-pick of the same commit in the 9.x branch, taken
from:

    https://github.com/liferay/liferay-js-themes-toolkit/pull/289

Key difference here is that `resolveDependency()` is quite different in
this branch but the argument still applies: we don't need the external
dep and can use the `paths` option to `require.resolve()` that has
existed since Node v8.9.

I also removed the unused `dirname` parameter to this function, because
no callsite is ever using it.

Test plan: `yarn test`

Related: https://github.com/liferay/liferay-js-themes-toolkit/issues/126