doc: remove Gulp entry from version table
At the time I wrote the table, I thought we were going to use Gulp v3 in
version 8 and Gulp v4 in version 9, but in the end we stayed with Gulp
v3 in both, so we should remove it from the table to avoid confusion. (I
already removed the reference to v4 in 7a6da6568894f3).