fix: failure to update after non-JS/CSS changes in watch mode (#291)
It seems that both "deploy:file" and "deploy:folder" are horribly
broken, which means that on changing a template file nothing gets
deployed.

If you look at the code, `fastDeploy()` looks for a non-existent
"../../temp" folder relative to the theme files, so Gulp doesn't put
anything there for the portal to pick up. I considered instead getting
the "appServerPath" from the "liferay-theme.json" file, which does
existing, and using its "temp" directory, but it is not clear what the
internal structure of it should be, nor whether it will have the effect
of reloading the module in the same way that the magical "deploy" folder
does. In short, I think that this might have worked once upon a time in
conjunction with the right Gogo Shell commands, but now our only choice
is to do a slow-but-correct full deploy. "slow" is relative here, and
may only be a few seconds in the typical case, and is certainly much
better than "fast and broken".

Test plan: `gulp watch`, make a change to
"src/templates/portal_normal.ptl", see rebuild and deploy occur, and see
it reflected in tomcat log. Now make a change to "src/css/_custom.scss"
and see it do a partial rebuild, and no output in tomcat log. In both
cases, see the browser window reload automatically.

Closes: https://github.com/liferay/liferay-js-themes-toolkit/issues/291