feat: look for Sass dependency in top-level node_modules (#72)
The idea is to make use of the new `--sass-include-paths`
setting which will be added to the liferay-theme-tasks in:

https://github.com/liferay/liferay-js-themes-toolkit/pull/306

so that we can get themes to look up Sass dependencies in the top-level
"node_modules" folder in the liferay-portal repo.

This is necessitated because of changes to where/how we use font-awesome
in the portal.

Closes: https://github.com/liferay/liferay-npm-tools/issues/72