feat: simplify release process (#50)
As you can see in the changes to CONTRIBUTING.md, the release process
can be made simpler by doing two things:

- Adding `.yarnrc` config files that set up the templates for the
  correct tag and release commit message formats.
- Using `"preversion"` scripts to automatically run the CI checks
  (formatting, linting, and testing) before cutting a release.

Note that we run the checks for *all* packages in the repo, which
is overkill, but it is the simplest thing to do. If a problem in
package A ever blocks you from publishing package B you can pass
`--ignore-scripts`, but you should probably just fix package A because
we want the CI to be green on every commit anyway. As a follow-up, we
should make it possible to run the tests independently within a given
repo.

Related: https://github.com/liferay/liferay-npm-tools/issues/50