feat: simplify release process (#50)
As you can see in the changes to CONTRIBUTING.md, the release process
can be made simpler by doing two things:

- Adding `.yarnrc` config files that set up the templates for the
  correct tag and release commit message formats.
- Using `"preversion"` scripts to automatically run the CI checks
  (formatting, linting, and testing) before cutting a release.

Note that we run the checks for *all* packages in the repo, which
is overkill, but it is the simplest thing to do. If a problem in
package A ever blocks you from publishing package B you can pass
`--ignore-scripts`, but you should probably just fix package A because
we want the CI to be green on every commit anyway. As a follow-up, we
should make it possible to run the tests independently within a given
repo.

There are still further steps we can take to make releasing even simpler
(see the related issue), but this is a good first step.

Test plan:

Prepare some pretend releases using different invocations of `yarn
version`, see the CI runs succeed, and then inspect output of
`git log -p` to see that created commits include the right messages and
tags.

    cd packages
    cd liferay-jest-junit-reporter
    yarn version --minor
    cd ../liferay-npm-scripts
    yarn version --patch
    cd ../liferay-npm-bundler-preset-liferay-dev
    yarn version
    git log -p
    git tag -l

Related: https://github.com/liferay/liferay-npm-tools/issues/50