fix: handle missing theme resources as 404s (#309, #310)
Given a theme without a thumbnail, we would try to request the image
like at a path like "test-theme/build/images/thumbnail.png". This would
log an "Error: ENOENT" to the console, and send an empty response back
to the browser.

With this commit, we check to see if the file exists and serve it up if
it does; otherwise we just proxy the request as usual and in that way
the client gets a proper 404 response instead.