feat: add "wepback" subcommand to liferay-npm-scripts
At the moment all this does is run "webpack", which will use the default
"webpack.config.js" found in the current directory.

I have a liferay-portal PR under development which does some shuffling
around of files in that repo in order to make this possible:

https://github.com/wincent/liferay-portal/pull/8

Next steps will include getting a usable "watch" command in place, and
exploring whether we can offer a meaningful default config here. We'll
also have the option of integrating the webpack build into the standard
"liferay-npm-scripts build" command: the idea there will could be to run
the webpack step if a config file is present; in order to make that
happen though, will need to make some further changes to the upstream
projects (for example, segments/segments-web is already using a custom
invocation of `babel`, so would need to be adapted).

Based on:

https://github.com/liferay/liferay-npm-tools/pull/78

which will need to be merged before this one.