feat: handle lightweight tags in changelog script
As noted in #334 we accidentally created a lightweight (ie.
non-annotated) tag in the AlloyEditor project, which lead the changelog
script to not "see" the previous release tag, seek to the one prior to
that, and thus include some duplicate entries in the changelog.

We can avoid that by passing the `--tags` switch to `git describe`, but
I'll also add a check in AlloyEditor to make sure we don't accidentally
create non-annotated tags in the future.

Closes: https://github.com/liferay/liferay-js-themes-toolkit/issues/334