feat: always clean up temporary Babel config files
As previously implemented, any error thrown while running a command
after setting up Babel would leave a dirty work tree:

    setBabelConfig(); // mutates work tree
    spawnSync('boom!'); // will throw an error
    removeBabelConfig(); // never runs

This can lead to files getting accidentally committed, like in:

https://github.com/liferay/liferay-portal/commit/bed7ddb7cfc4f724079

So, make this "foolproof" by replacing those `set` and `remove`
utilities with a single `withBabelConfig()` helper that always cleans up
after itself because it uses a `finally` block.

(Not 100% "foolproof" because you can still screw up if you pass an
asynchronous callback; if that ever happens then it will be proof that
entropy will one day be the end of us all.)

Adding this because I am about to add a third location where we do this
Babel set-up, and I don't want the old pattern to extend any further.