feat: add Babel merge strategy to get-merged-config.js
As described in the related issue, our merge implementation is using the
pre-2.0 version of the array merge algorithm in the "deepmerge" package.
They removed it because it was a "complicated thing".

In the case of merging Babel configs it verily breaks them, because you
can specify plugins and presets as either `'name'` or `['name',
options]`. Here's a reduced example of some surprising behavior:

- merging ['a', 'b', ['c']]
- into ['d', 'e']
- yields: ['d', 'e', ['c'], 'b']

ie. 'a' disappears into the void.

See the test cases included in this commit for a real-life example taken
from segments-web. The old algorithm would cause the
`@babel/plugin-proposal-class-properties` plugin to vanish.

The implementation here doesn't intend to handle all possible edge cases
(for example, it doesn't try to match plugins that have been written in
different forms such as "@babel/plugin-x" and "x"), but it covers the
ones we are most likely to see in liferay-portal and it is well tested.

I chose to write a narrow implementation like this rather than pull in a
third-party dependency like "babel-merge", because I like to keep our
dependency footprint small and we actually kind of want to discourage
complicated Babel overrides in liferay-portal in any case, so it's in
our interests to have something that isn't too flexible.

Closes: https://github.com/liferay/liferay-npm-tools/issues/83